Balance factor fixed

+++ 14-binary_tree_balance.c
@@ -24,9 +24,9 @@ int binary_tree_balance(const binary_tree_t *tree)
  * @tree: pointer to root node
  * Return: calculated height
  */
-int binary_tree_height(const binary_tree_t *tree)
+size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int left_height, right_height;
+	size_t left_height, right_height;
 
 	if (tree == NULL)
 	{
